datahub/pyproject.toml
[tool.poetry]
name = "oaf-datahub"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
acryl-datahub = "^0.8.0"
acryl-datahub-actions = "^0.0.8"
requests = "^2.28.1"
gql = "^3.4.0"
requests-toolbelt = "^0.10.1" # Required by gql but not automatically installed for some reason

[tool.poetry.dev-dependencies]
mypy = "^0.991"
pylint = "^2.15.6"
types-requests = "^2.28.11"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
